Summit-RCM: AT Interface: Standardize/Cleanup AT Commands
Bug 24986

+++ summit_rcm/at_interface/commands/fips_command.py
@@ -1,10 +1,25 @@
+"""
+File that consists of the Fips Command Functionality
+"""
+from syslog import LOG_ERR, syslog
 from typing import List, Tuple
+from enum import IntEnum
 from summit_rcm.at_interface.commands.command import Command
-from summit_rcm.services.fips_service import FipsService
+from summit_rcm.services.fips_service import FipsService, FipsUnsupportedError
 from summit_rcm.services.system_service import SystemService
 
 
+class States(IntEnum):
+    fips = 0
+    fips_wifi = 1
+    unset = 2
+
+
 class FipsCommand(Command):
+    """
+    AT Command to get/set fips state
+    """
+
     NAME: str = "Fips"
     SIGNATURE: str = "at+fips"
     VALID_NUM_PARAMS: List[int] = [1, 2]
@@ -14,19 +29,23 @@ class FipsCommand(Command):
     async def execute(params: str) -> Tuple[bool, str]:
         (valid, params_dict) = FipsCommand.parse_params(params)
         if not valid:
-            return (
-                True,
-                f"\r\nInvalid Parameters: See Usage - {FipsCommand.SIGNATURE}?\r\n",
-            )
+            syslog(LOG_ERR, "Invalid Parameters")
+            return (True, "\r\nERROR\r\n")
         try:
             if params_dict["state"]:
-                fips_str = str(await FipsService().set_fips_state(params_dict["state"]))
+                if await FipsService().get_fips_state() == "unsupported":
+                    raise FipsUnsupportedError("Fips Unsupported")
+                success = await FipsService().set_fips_state(params_dict["state"])
+                if not success:
+                    return (True, "\r\nERROR\r\n")
                 if params_dict["reboot"]:
                     await SystemService().set_power_state("reboot")
+                return (True, "\r\nOK\r\n")
             else:
                 fips_str = await FipsService().get_fips_state()
-            return (True, f"\r\n+FIPS:{fips_str}\r\nOK\r\n")
-        except Exception:
+            return (True, f"\r\n+FIPS: {fips_str}\r\nOK\r\n")
+        except Exception as exception:
+            syslog(LOG_ERR, f"Error setting/getting FIPS state: {str(exception)}")
             return (True, "\r\nERROR\r\n")
 
     @staticmethod
@@ -36,15 +55,17 @@ class FipsCommand(Command):
         params_list = params.split(",")
         given_num_param = len(params_list)
         valid &= given_num_param in FipsCommand.VALID_NUM_PARAMS
-        if valid and params_list[0].lower() in FipsCommand.VALID_STATE_VAL:
-            params_dict["state"] = params_list[0]
-        elif params_list[0].lower():
+        if not valid:
+            return (False, {})
+        try:
+            params_dict["state"] = (
+                States(int(params_list[0])).name if params_list[0] else ""
+            )
+            params_dict["reboot"] = (
+                bool(int(params_list[1])) if len(params_list) > 1 else False
+            )
+        except ValueError:
             valid = False
-        else:
-            params_dict["state"] = ""
-        params_dict["reboot"] = (
-            1 if (len(params_list) > 1 and params_list[1] == "1") else 0
-        )
         return (valid, params_dict)
 
     @staticmethod
